Docker CLI プラグイン (docker-torii) の実装
概要: docker run のたびに .env ファイルをディスク上に作りたくないという課題

理由:
- docker compose や docker run で暗号化済み環境変数を安全に渡す手段がなかった
- --env フラグは docker inspect やプロセスリストから平文が見える

変更内容:
- src/lib.rs を新規作成し、既存コードをライブラリ化
- cmd_* 関数群を src/commands.rs に移動（main.rs はディスパッチのみ）
- docker-torii バイナリを追加（Docker CLI プラグイン）
  - docker-cli-plugin-metadata でメタデータ応答
  - run サブコマンド: 復号 → tmpfs に .env 書き出し → bind mount で docker run
  - コンテナ終了後に .env を zeroize → tmpfs 削除
  - Linux では tmpfs マウントを試行（失敗時は通常 tempdir にフォールバック）
- .env ファイルは 0400、tmpfs ディレクトリは 0700 で保護
