+++ src/lib.rs
@@ -1,0 +1,8 @@
+pub mod cli;
+pub mod commands;
+pub mod crypto;
+pub mod db;
+pub mod error;
+pub mod fuse_fs;
+pub mod logger;
+pub mod tui;
